Fixing Task3, Task5, Task6, Task7

+++ Tyuiu.VasilevAA.Sprint1.Task5.V2.Lib/DataService.cs
@@ -10,7 +10,7 @@ namespace Tyuiu.VasilevAA.Sprint1.Task5.V2.Lib
 {
     public class DataService : ISprint1Task5V2
     {
-        public double FahrenheitToCelsius(double tempFahrenheit)
+        public int FahrenheitToCelsius(double tempFahrenheit)
         {
             return Convert.ToInt32((tempFahrenheit - 32) * 5 / 9);
         }
